Alterado o backend para ter o controle do saldo e criado o endpoint transacoes para poder movimentar entre contas (#71)
Alterado o backend para ter o controle do saldo e criado o endpoint transacoes para poder movimentar entre contas

+++ d3-finance-api/d3_finance_api/src/api/v1/endpoints/receita_controller.py
@@ -3,7 +3,7 @@ from fastapi import HTTPException, Depends, status, Response
 from sqlalchemy.orm import Session
 from src.app import router
 from src.database.database import SessionLocal
-from src.database.models import Receitas
+from src.database.models import Receitas, Contas
 from src.api.tags import Tag
 from src.schemas.receita_schemas import ReceitaCreate, ReceitaUpdate, ReceitaResponse
 
@@ -31,7 +31,6 @@ def get_receitas(db: Session = Depends(get_db)):
     return [ReceitaResponse(
         id=receita.id,
         categoria=receita.categoria,
-        nome_receita=receita.nome_receita,
         valor_recebido=receita.valor_recebido,
         data_recebimento=receita.data_recebimento,
         descricao=receita.descricao,
@@ -55,7 +54,6 @@ def get_receita_by_id(receitas_id: int, db: Session = Depends(get_db)):
     return ReceitaResponse(
         id=receita.id,
         categoria=receita.categoria,
-        nome_receita=receita.nome_receita,
         valor_recebido=receita.valor_recebido,
         data_recebimento=receita.data_recebimento,
         descricao=receita.descricao,
@@ -70,10 +68,13 @@ def get_receita_by_id(receitas_id: int, db: Session = Depends(get_db)):
     path=CADASTRO_RECEITAS, response_model=ReceitaResponse, tags=[Tag.Receitas.name]
 )
 def create_receita(receita: ReceitaCreate, db: Session = Depends(get_db)):  
-  
+    # Verifica se conta existe
+    conta = db.query(Contas).filter(Contas.id == receita.conta_id).first()
+    if not conta:
+        raise HTTPException(status_code=404, detail="Conta não encontrada")
+    
     db_receita = Receitas(
         categoria=receita.categoria,
-        nome_receita=receita.nome_receita,
         valor_recebido=receita.valor_recebido,
         data_recebimento=receita.data_recebimento,
         forma_recebimento=receita.forma_recebimento,
@@ -81,24 +82,27 @@ def create_receita(receita: ReceitaCreate, db: Session = Depends(get_db)):
         conta_id=receita.conta_id,
     )
 
-    db.add(db_receita)  
-    db.commit()  
-    db.refresh(db_receita)  
+    db.add(db_receita)
+
+    # Atualiza saldo da conta
+    conta.saldo += receita.valor_recebido
+
+    db.commit()
+    db.refresh(db_receita)
 
     return ReceitaResponse(
         id=db_receita.id,
         categoria=db_receita.categoria,
-        nome_receita=db_receita.nome_receita,
         valor_recebido=db_receita.valor_recebido,
         data_recebimento=db_receita.data_recebimento,
         descricao=db_receita.descricao,
         forma_recebimento=db_receita.forma_recebimento,
-        conta_id=receita.conta_id,
+        conta_id=db_receita.conta_id,
         data_criacao=db_receita.data_criacao,
         data_alteracao=db_receita.data_alteracao,
     )
 
-# PUT - Atualizar receita
+
 @router.put(
     path=ATUALIZAR_RECEITAS, response_model=ReceitaResponse, tags=[Tag.Receitas.name]
 )
@@ -108,10 +112,21 @@ def update_receita(receitas_id: int, receita_update: ReceitaUpdate, db: Session 
     if not receita:
         raise HTTPException(status_code=404, detail="Receita não encontrada")
 
+    conta = db.query(Contas).filter(Contas.id == receita.conta_id).first()
+    if not conta:
+        raise HTTPException(status_code=404, detail="Conta não encontrada")
+
+    # Ajusta o saldo da conta:
+    # Subtrai o valor antigo da receita do saldo da conta
+    conta.saldo -= receita.valor_recebido
+
     # Atualiza os campos da receita
     for field, value in receita_update.__dict__.items():
         if value is not None:
             setattr(receita, field, value)
+
+    # Soma o novo valor da receita no saldo da conta
+    conta.saldo += receita.valor_recebido
 
     db.commit()
     db.refresh(receita)
@@ -119,7 +134,6 @@ def update_receita(receitas_id: int, receita_update: ReceitaUpdate, db: Session 
     return ReceitaResponse(
         id=receita.id,
         categoria=receita.categoria,
-        nome_receita=receita.nome_receita,
         valor_recebido=receita.valor_recebido,
         data_recebimento=receita.data_recebimento,
         descricao=receita.descricao,
@@ -138,8 +152,15 @@ def delete_receita(receitas_id: int, db: Session = Depends(get_db)):
 
     if not receita:
         raise HTTPException(status_code=404, detail="Receita não encontrada")
-    
+
+    conta = db.query(Contas).filter(Contas.id == receita.conta_id).first()
+    if not conta:
+        raise HTTPException(status_code=404, detail="Conta não encontrada")
+
+    # Subtrai o valor da receita do saldo da conta
+    conta.saldo -= receita.valor_recebido
+
     db.delete(receita)
     db.commit()
-    
+
     return Response(status_code=status.HTTP_204_NO_CONTENT)
